+++ app/src/main/AndroidManifest.xml
@@ -28,25 +28,19 @@
             android:label="@string/title_activity_vue_ajout_evenement"
             android:theme="@style/AppTheme.NoActionBar" />
         <activity android:name=".VueCalendrierCache" />
-        <activity android:name=".VueListeCachee" />
+        <activity android:name=".VueListeCachee"></activity>
         <activity android:name=".VueAjoutContactPrive">
             <intent-filter>
-                <action android:name="com.company.package.FOO" />
-
+                <action android:name="com.company.package.FOO"/>
                 <category android:name="android.intent.category.DEFAULT" />
             </intent-filter>
         </activity>
-        <activity android:name=".VueSuppressionContact">
+        <activity android:name=".VueModificationSuppressionContact">
             <intent-filter>
-                <action android:name="com.company.package.FOO" />
-
+                <action android:name="com.company.package.FOO"/>
                 <category android:name="android.intent.category.DEFAULT" />
             </intent-filter>
         </activity>
-        <activity android:name=".VueModifierEvenement" />
-        <activity android:name=".VueAjoutEvenementCache" />
-        <activity android:name=".VueConsulterEvenementCache" />
-        <activity android:name=".VueListeContactsCache"></activity>
     </application>
 
 </manifest>
