--- app/src/main/res/layout/activity_vue_suppression_contact.xml
@@ -1,195 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-
-    <TextView
-        android:id="@+id/textView5"
-        android:layout_width="413dp"
-        android:layout_height="32dp"
-        android:background="@color/colorAccent"
-        android:text="Modification de contact"
-        android:textAlignment="center"
-        android:textColor="@color/colorText"
-        tools:layout_editor_absoluteX="94dp"
-        tools:layout_editor_absoluteY="5dp" />
-
-    <EditText
-        android:id="@+id/editText2"
-        android:layout_width="417dp"
-        android:layout_height="33dp"
-        android:background="@color/colorAccent"
-        android:ems="10"
-        android:inputType="textPersonName"
-        android:text="Martin"
-        android:textColor="@color/colorText"
-        tools:layout_editor_absoluteX="92dp"
-        tools:layout_editor_absoluteY="315dp" />
-
-    <EditText
-        android:id="@+id/editText3"
-        android:layout_width="417dp"
-        android:layout_height="33dp"
-        android:background="@color/colorAccent"
-        android:ems="10"
-        android:inputType="textPersonName"
-        android:text="Isabelle"
-        android:textColor="@color/colorText"
-        tools:layout_editor_absoluteX="92dp"
-        tools:layout_editor_absoluteY="379dp" />
-
-    <EditText
-        android:id="@+id/editText5"
-        style="?attr/editTextStyle"
-        android:layout_width="417dp"
-        android:layout_height="33dp"
-        android:background="@color/colorAccent"
-        android:ems="10"
-        android:inputType="phone"
-        android:text="06 05 06 05 06"
-        android:textColor="@color/colorText"
-        tools:layout_editor_absoluteX="92dp"
-        tools:layout_editor_absoluteY="440dp" />
-
-    <EditText
-        android:id="@+id/editText6"
-        android:layout_width="417dp"
-        android:layout_height="33dp"
-        android:background="@color/colorAccent"
-        android:ems="10"
-        android:inputType="textPostalAddress"
-        android:text="02 rue du janvier"
-        android:textColor="@color/colorText"
-        tools:layout_editor_absoluteX="92dp"
-        tools:layout_editor_absoluteY="497dp" />
-
-    <EditText
-        android:id="@+id/editText7"
-        android:layout_width="417dp"
-        android:layout_height="33dp"
-        android:background="@color/colorAccent"
-        android:ems="10"
-        android:inputType="textEmailAddress"
-        android:text="i.martin@gmail.com"
-        android:textColor="@color/colorText"
-        tools:layout_editor_absoluteX="92dp"
-        tools:layout_editor_absoluteY="555dp" />
-
-    <EditText
-        android:id="@+id/editText9"
-        android:layout_width="353dp"
-        android:layout_height="32dp"
-        android:background="@color/colorAccent"
-        android:ems="10"
-        android:inputType="textPersonName"
-        android:text="Une énigme"
-        android:textColor="@color/colorText"
-        tools:layout_editor_absoluteX="92dp"
-        tools:layout_editor_absoluteY="623dp" />
-
-    <NumberPicker
-        android:id="@+id/numberPicker"
-        android:layout_width="41dp"
-        android:layout_height="47dp"
-        android:background="@drawable/linearborder"
-        tools:layout_editor_absoluteX="466dp"
-        tools:layout_editor_absoluteY="614dp" />
-
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/floatingActionButton2"
-        android:layout_width="73dp"
-        android:layout_height="68dp"
-        android:clickable="true"
-        app:srcCompat="@android:drawable/ic_menu_add"
-        tools:layout_editor_absoluteX="514dp"
-        tools:layout_editor_absoluteY="718dp" />
-
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/floatingActionButton3"
-        android:layout_width="72dp"
-        android:layout_height="68dp"
-        android:clickable="true"
-        app:srcCompat="@android:drawable/ic_menu_delete"
-        tools:layout_editor_absoluteX="430dp"
-        tools:layout_editor_absoluteY="718dp" />
-
-    <TextView
-        android:id="@+id/AlertTextView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/floatingActionButton3"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="10dp"
-        android:text="Êtes-vous sûr de vouloir supprimer ce contact ? "
-        android:textSize="20sp"
-        android:visibility="invisible"
-        tools:layout_editor_absoluteX="94dp"
-        tools:layout_editor_absoluteY="295dp" />
-
-
-    <ImageView
-        android:id="@+id/imageView"
-        android:layout_width="414dp"
-        android:layout_height="241dp"
-        app:srcCompat="@drawable/woman"
-        tools:layout_editor_absoluteX="94dp"
-        tools:layout_editor_absoluteY="54dp" />
-
-    <ImageButton
-        android:id="@+id/imageButton"
-        android:layout_width="38dp"
-        android:layout_height="37dp"
-        android:background="@android:color/white"
-        android:tint="@color/colorAccent"
-        app:srcCompat="@android:drawable/ic_menu_myplaces"
-        tools:layout_editor_absoluteX="38dp"
-        tools:layout_editor_absoluteY="314dp" />
-
-    <ImageButton
-        android:id="@+id/imageButton5"
-        android:layout_width="38dp"
-        android:layout_height="37dp"
-        android:background="@android:color/white"
-        android:tint="@color/colorAccent"
-        app:srcCompat="@android:drawable/ic_dialog_map"
-        tools:layout_editor_absoluteX="38dp"
-        tools:layout_editor_absoluteY="496dp" />
-
-    <ImageButton
-        android:id="@+id/imageButton6"
-        android:layout_width="38dp"
-        android:layout_height="37dp"
-        android:background="@android:color/white"
-        android:tint="@color/colorAccent"
-        app:srcCompat="@android:drawable/ic_dialog_email"
-        tools:layout_editor_absoluteX="38dp"
-        tools:layout_editor_absoluteY="554dp" />
-
-    <ImageButton
-        android:id="@+id/imageButton7"
-        android:layout_width="38dp"
-        android:layout_height="37dp"
-        android:background="@android:color/white"
-        android:tint="@color/colorAccent"
-        app:srcCompat="@android:drawable/ic_menu_call"
-        tools:layout_editor_absoluteX="38dp"
-        tools:layout_editor_absoluteY="439dp" />
-
-    <ImageButton
-        android:id="@+id/imageButton8"
-        android:layout_width="38dp"
-        android:layout_height="37dp"
-        android:background="@android:color/white"
-        android:tint="@color/colorAccent"
-        app:srcCompat="@android:drawable/ic_menu_help"
-        tools:layout_editor_absoluteX="38dp"
-        tools:layout_editor_absoluteY="622dp" />
-
-
-
-
-</android.support.constraint.ConstraintLayout>
